move trends to dedicated package
reivse/add comments and order of funcs/definitions
