+++ engine/dynamicdp.go
@@ -131,7 +131,7 @@ func (dDP *dynamicDP) fieldAsInterface(fldPath []string) (val any, err error) {
 
 	case utils.MetaTrends:
 		//sample of fieldName : ~*trends.TrendID.Metrics.*acd.Value
-		var trendSum TrendSummary
+		var trendSum utils.TrendSummary
 		if err := connMgr.Call(context.TODO(), dDP.trdConns, utils.TrendSv1GetTrendSummary, &utils.TenantIDWithAPIOpts{TenantID: &utils.TenantID{Tenant: dDP.tenant, ID: fldPath[1]}}, &trendSum); err != nil {
 			return nil, err
 		}
