+++ engine/storage_interface.go
@@ -70,11 +70,11 @@ type DataDB interface {
 	SetRankingDrv(ctx *context.Context, rn *Ranking) (err error)
 	GetRankingDrv(ctx *context.Context, tenant string, id string) (sq *Ranking, err error)
 	RemoveRankingDrv(ctx *context.Context, tenant string, id string) (err error)
-	SetTrendProfileDrv(ctx *context.Context, sq *TrendProfile) (err error)
-	GetTrendProfileDrv(ctx *context.Context, tenant string, id string) (sq *TrendProfile, err error)
+	SetTrendProfileDrv(ctx *context.Context, sq *utils.TrendProfile) (err error)
+	GetTrendProfileDrv(ctx *context.Context, tenant string, id string) (sq *utils.TrendProfile, err error)
 	RemTrendProfileDrv(ctx *context.Context, tenant string, id string) (err error)
-	GetTrendDrv(ctx *context.Context, tenant string, id string) (*Trend, error)
-	SetTrendDrv(ctx *context.Context, tr *Trend) error
+	GetTrendDrv(ctx *context.Context, tenant string, id string) (*utils.Trend, error)
+	SetTrendDrv(ctx *context.Context, tr *utils.Trend) error
 	RemoveTrendDrv(ctx *context.Context, tenant string, id string) error
 	GetFilterDrv(ctx *context.Context, tnt string, id string) (*Filter, error)
 	SetFilterDrv(ctx *context.Context, f *Filter) error
