+++ loaders/loader.go
@@ -110,7 +110,7 @@ func setToDB(ctx *context.Context, dm *engine.DataManager, lType string, data pr
 	case utils.MetaAccounts:
 		return dm.SetAccount(ctx, data.(*utils.Account), withIndex)
 	case utils.MetaTrends:
-		return dm.SetTrendProfile(ctx, data.(*engine.TrendProfile))
+		return dm.SetTrendProfile(ctx, data.(*utils.TrendProfile))
 	case utils.MetaRankings:
 		return dm.SetRankingProfile(ctx, data.(*engine.RankingProfile))
 	}
